requirements.txt
pytest==9.0.2
playwright==1.58.0
requests==2.32.5
pytest-playwright==0.7.2
pytest-html==4.2.0
pytest-xdist==3.8.0
python-dotenv==1.2.1